Improve components defined for env-details page
